requirements.txt
Django==5.2.4
djangorestframework==3.14.0
psycopg2-binary==2.9.7
requests==2.31.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0